Rename raw-results-table.tsx => RawTable.tsx
