+++ extensions/ql-vscode/src/view/results/RawTable.tsx
