+++ extensions/ql-vscode/src/view/results/ResultTable.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { AlertTable } from "./AlertTable";
 import { Graph } from "./Graph";
-import { RawTable } from "./raw-results-table";
+import { RawTable } from "./RawTable";
 import { ResultTableProps } from "./result-table-utils";
 
 export function ResultTable(props: ResultTableProps) {
